using double

+++ Core/include/Acts/Vertexing/HoughVertexFinder.hpp
@@ -42,7 +42,7 @@ namespace Acts {
 /// 4. Makes a projection to the Z axis and finds a peak - that is the vertex
 /// position
 /// 5. Repeats 2-4 if necessary
-
+///
 template <typename spacepoint_t>
 class HoughVertexFinder {
  public:
@@ -54,8 +54,8 @@ class HoughVertexFinder {
 
     /// Minimum and maximum ranges in |eta|; the |eta| will not be
     /// set outside these bounds even if targetSPs is not reached
-    float minAbsEta = 0.3f;
-    float maxAbsEta = 4.0f;
+    double minAbsEta = 0.3f;
+    double maxAbsEta = 4.0f;
 
     /// Minimum number of hits in Hough plane to consider
     /// the cell to contain a track
@@ -70,20 +70,20 @@ class HoughVertexFinder {
     /// of the measurements.
     /// The first |eta| range starts at 0., the others start at
     /// the endpoint of the previous range.
-    std::vector<float> absEtaRanges{2.f, 4.f};
+    std::vector<double> absEtaRanges{2., 4.};
     /// The amount of measurements in the |eta| range expressed
     /// as a fraction of all measurements within the whole |eta| range.
     /// Measurements are assumed to be distributed uniformly with
     /// the |eta| range.
-    std::vector<float> absEtaFractions{0.4f, 0.6f};
+    std::vector<double> absEtaFractions{0.4f, 0.6f};
 
     /// The algorithm starts peak searching considering wide range
     /// in Z and then iteratively constrain itself to narrower ranges
     /// around previously found peak. At the same time it may adapt
     /// other Hough-image parameters in each step.
-    std::vector<float> rangeIterZ{200.f * UnitConstants::mm,
-                                  30.f * UnitConstants::mm,
-                                  16.f * UnitConstants::mm};
+    std::vector<double> rangeIterZ{200. * UnitConstants::mm,
+                                   30. * UnitConstants::mm,
+                                   16. * UnitConstants::mm};
     /// Number of bins along z-axis of the Hough image for each iteration
     std::vector<std::uint32_t> nBinsZIterZ{800, 180, 80};
     /// Number of bins along cot(theta)-axis of the Hough image for
@@ -95,14 +95,14 @@ class HoughVertexFinder {
     /// number of bins can be smaller than stated in "nBinsCotThetaIterZ".
     /// For every magnitude (in natural logarithm) below targetSPs, the number
     /// of bins in cot(theta) will decrease by this factor.
-    float binsCotThetaDecrease = 1.35f;
+    double binsCotThetaDecrease = 1.35f;
 
     /// Width of the peak when estimating vertex position
     std::uint32_t peakWidth = 3;
 
     /// Default position of the vertex in X, Y, and Z coordinates
-    Vector3 defVtxPosition{0.f * UnitConstants::mm, 0.f * UnitConstants::mm,
-                           0.f * UnitConstants::mm};
+    Vector3 defVtxPosition{0. * UnitConstants::mm, 0. * UnitConstants::mm,
+                           0. * UnitConstants::mm};
   };
 
   /// Const access to the config
@@ -141,16 +141,16 @@ class HoughVertexFinder {
   /// @return Position of the vertex in (X,Y,Z)
   Acts::Result<Acts::Vector3> findHoughVertex(
       const std::vector<spacepoint_t>& spacepoints, const Acts::Vector3& vtxOld,
-      float rangeZ, std::uint32_t numZBins, float minCotTheta,
-      float maxCotTheta, std::uint32_t numCotThetaBins) const;
+      double rangeZ, std::uint32_t numZBins, double minCotTheta,
+      double maxCotTheta, std::uint32_t numCotThetaBins) const;
 
   /// @brief Finds the peak in the Z axis projection of the Hough space
   /// @param houghZProjection Hough space projection after the cleaning procedure
   /// @param vtxZPositions Bins position in the Hough space projection
   /// @return Position of the peak
-  Acts::Result<float> findHoughPeak(
+  Acts::Result<double> findHoughPeak(
       const std::vector<std::uint32_t>& houghZProjection,
-      const std::vector<float>& vtxZPositions) const;
+      const std::vector<double>& vtxZPositions) const;
 
   /// Logging instance
   std::unique_ptr<const Logger> m_logger;
